+++ modal/dict.py
@@ -2,25 +2,32 @@
 from collections.abc import AsyncIterator, Mapping
 from dataclasses import dataclass
 from datetime import datetime
-from typing import Any, Optional
+from typing import Any, Optional, Union
 
 from google.protobuf.message import Message
 from grpclib import GRPCError
+from synchronicity import classproperty
 from synchronicity.async_wrap import asynccontextmanager
 
 from modal_proto import api_pb2
 
-from ._object import EPHEMERAL_OBJECT_HEARTBEAT_SLEEP, _get_environment_name, _Object, live_method, live_method_gen
+from ._object import (
+    EPHEMERAL_OBJECT_HEARTBEAT_SLEEP,
+    _get_environment_name,
+    _Object,
+    live_method,
+    live_method_gen,
+)
 from ._resolver import Resolver
 from ._serialization import deserialize, serialize
 from ._utils.async_utils import TaskContext, synchronize_api
 from ._utils.deprecation import deprecation_warning, warn_if_passing_namespace
 from ._utils.grpc_utils import retry_transient_errors
 from ._utils.name_utils import check_object_name
-from ._utils.time_utils import timestamp_to_localized_dt
+from ._utils.time_utils import as_timestamp, timestamp_to_localized_dt
 from .client import _Client
 from .config import logger
-from .exception import RequestSizeError
+from .exception import InvalidError, NotFoundError, RequestSizeError
 
 
 def _serialize_dict(data):
@@ -29,7 +36,7 @@ def _serialize_dict(data):
 
 @dataclass
 class DictInfo:
-    """Information about the Dict object."""
+    """Information about a Dict object."""
 
     # This dataclass should be limited to information that is unchanging over the lifetime of the Dict,
     # since it is transmitted from the server when the object is hydrated and could be stale when accessed.
@@ -37,6 +44,102 @@ class DictInfo:
     name: Optional[str]
     created_at: datetime
     created_by: Optional[str]
+
+
+class _DictManager:
+    """Namespace with methods for managing named Dict objects."""
+
+    @staticmethod
+    async def list(
+        *,
+        max_objects: Optional[int] = None,  # Limit results to this size
+        created_before: Optional[Union[datetime, str]] = None,  # Limit based on creation date
+        environment_name: str = "",  # Uses active environment if not specified
+        client: Optional[_Client] = None,  # Optional client with Modal credentials
+    ) -> list["_Dict"]:
+        """Return a list of hydrated Dict objects.
+
+        **Examples:**
+
+        ```python
+        dicts = modal.Dict.objects.list()
+        print([d.name for d in dicts])
+        ```
+
+        Dicts will be retreived from the active environment, or another one can be specified:
+
+        ```python notest
+        dev_dicts = modal.Dict.objects.list(environment_name="dev")
+        ```
+
+        By default, all named Dict are returned, newest to oldest. It's also possible to limit the
+        number of results and to filter by creation date:
+
+        ```python
+        dicts = modal.Dict.objects.list(max_objects=10, created_before="2025-01-01")
+        ```
+
+        """
+        client = await _Client.from_env() if client is None else client
+        if max_objects is not None and max_objects < 0:
+            raise InvalidError("max_objects cannot be negative")
+
+        items: list[api_pb2.DictListResponse.DictInfo] = []
+
+        async def retrieve_page(created_before: float) -> bool:
+            max_page_size = 100 if max_objects is None else min(100, max_objects - len(items))
+            pagination = api_pb2.ListPagination(max_objects=max_page_size, created_before=created_before)
+            req = api_pb2.DictListRequest(environment_name=environment_name, pagination=pagination)
+            resp = await retry_transient_errors(client.stub.DictList, req)
+            items.extend(resp.dicts)
+            finished = (len(resp.dicts) < max_page_size) or (max_objects is not None and len(items) >= max_objects)
+            return finished
+
+        finished = await retrieve_page(as_timestamp(created_before))
+        while True:
+            if finished:
+                break
+            finished = await retrieve_page(items[-1].metadata.creation_info.created_at)
+
+        dicts = [_Dict._new_hydrated(item.dict_id, client, item.metadata, is_another_app=True) for item in items]
+        return dicts[:max_objects] if max_objects is not None else dicts
+
+    @staticmethod
+    async def delete(
+        name: str,  # Name of the Dict to delete
+        *,
+        allow_missing: bool = False,  # If True, don't raise an error if the Dict doesn't exist
+        environment_name: Optional[str] = None,  # Uses active environment if not specified
+        client: Optional[_Client] = None,  # Optional client with Modal credentials
+    ):
+        """Delete a named Dict.
+
+        Warning: This deletes an *entire Dict*, not just a specific key.
+        Deletion is irreversible and will affect any Apps currently using the Dict.
+
+        **Examples:**
+
+        ```python notest
+        await modal.Dict.objects.delete("my-dict")
+        ```
+
+        Dicts will be deleted from the active environment, or another one can be specified:
+
+        ```python notest
+        await modal.Dict.objects.delete("my-dict", environment_name="dev")
+        ```
+        """
+        try:
+            obj = await _Dict.from_name(name, environment_name=environment_name).hydrate(client)
+        except NotFoundError:
+            if not allow_missing:
+                raise
+        else:
+            req = api_pb2.DictDeleteRequest(dict_id=obj.object_id)
+            await retry_transient_errors(obj._client.stub.DictDelete, req)
+
+
+DictManager = synchronize_api(_DictManager)
 
 
 class _Dict(_Object, type_prefix="di"):
@@ -89,6 +192,10 @@ class _Dict(_Object, type_prefix="di"):
         raise RuntimeError(
             "`Dict(...)` constructor is not allowed. Please use `Dict.from_name` or `Dict.ephemeral` instead"
         )
+
+    @classproperty
+    def objects(cls) -> _DictManager:
+        return _DictManager
 
     @property
     def name(self) -> Optional[str]:
@@ -238,9 +345,18 @@ class _Dict(_Object, type_prefix="di"):
         client: Optional[_Client] = None,
         environment_name: Optional[str] = None,
     ):
-        obj = await _Dict.from_name(name, environment_name=environment_name).hydrate(client)
-        req = api_pb2.DictDeleteRequest(dict_id=obj.object_id)
-        await retry_transient_errors(obj._client.stub.DictDelete, req)
+        """mdmd:hidden
+        Delete a named Dict object.
+
+        Warning: This deletes an *entire Dict*, not just a specific key.
+        Deletion is irreversible and will affect any Apps currently using the Dict.
+
+        DEPRECATED: This method is deprecated; we recommend using `modal.Dict.objects.delete` instead.
+        """
+        deprecation_warning(
+            (2025, 8, 6), "`modal.Dict.delete` is deprecated; we recommend using `modal.Dict.objects.delete` instead."
+        )
+        await _Dict.objects.delete(name, environment_name=environment_name, client=client)
 
     @live_method
     async def info(self) -> DictInfo: