+++ test/dict_test.py
@@ -1,5 +1,6 @@
 # Copyright Modal Labs 2022
 import pytest
+import sys
 import time
 
 from modal import Dict
@@ -36,9 +37,10 @@ def test_dict_named(servicer, client):
     d["foo"] = None
     assert d["foo"] is None
 
-    Dict.delete("my-amazing-dict", client=client)
+    Dict.objects.delete("my-amazing-dict", client=client)
     with pytest.raises(NotFoundError):
         Dict.from_name("my-amazing-dict").hydrate(client)
+    Dict.objects.delete("my-amazing-dict", client=client, allow_missing=True)
 
 
 def test_dict_ephemeral(servicer, client):
@@ -98,3 +100,20 @@ def test_dict_namespace_deprecated(servicer, client):
     # Filter out any unrelated warnings
     namespace_warnings = [w for w in record if "namespace" in str(w.message).lower()]
     assert len(namespace_warnings) == 0
+
+
+def test_dict_list(servicer, client):
+    for i in range(5):
+        Dict.from_name(f"test-dict-{i}", create_if_missing=True).hydrate(client)
+    if sys.platform == "win32":
+        time.sleep(1 / 32)
+
+    print(servicer.deployed_dicts)
+
+    dict_list = Dict.objects.list(client=client)
+    assert len(dict_list) == 5
+    assert all(d.name.startswith("test-dict-") for d in dict_list)
+    assert all(d.info().created_by == servicer.default_username for d in dict_list)
+
+    dict_list = Dict.objects.list(max_objects=2, client=client)
+    assert len(dict_list) == 2