+++ test/volume_test.py
@@ -451,9 +451,8 @@ async def test_volume_copy_2(client, tmp_path, servicer, version):
     assert returned_file_data[Path("test_dir/file2.txt")].data == b"test copy"
 
 
-@pytest.mark.parametrize("delete_as_instance_method", [True, False])
 @pytest.mark.parametrize("version", VERSIONS)
-def test_persisted(servicer, client, delete_as_instance_method, version):
+def test_from_name(servicer, client, version):
     # Lookup should fail since it doesn't exist
     with pytest.raises(NotFoundError):
         modal.Volume.from_name("xyz", version=version).hydrate(client)
@@ -464,12 +463,11 @@ def test_persisted(servicer, client, delete_as_instance_method, version):
     # Lookup should succeed now
     modal.Volume.from_name("xyz", version=version).hydrate(client)
 
-    # Delete it
-    modal.Volume.delete("xyz", client=client)
-
+    modal.Volume.objects.delete("xyz", client=client)
     # Lookup should fail again
     with pytest.raises(NotFoundError):
         modal.Volume.from_name("xyz", version=version).hydrate(client)
+    modal.Volume.objects.delete("xyz", client=client, allow_missing=True)
 
 
 def test_ephemeral(servicer, client):
@@ -575,3 +573,18 @@ def test_remove_file_not_found(set_env_client):
     vol = modal.Volume.from_name("my_vol", create_if_missing=True)
     with pytest.raises(FileNotFoundError):
         vol.remove_file("a")
+
+
+def test_volume_list(servicer, client):
+    for i in range(5):
+        modal.Volume.from_name(f"test-volume-{i}", create_if_missing=True).hydrate(client)
+    if sys.platform == "win32":
+        time.sleep(1 / 32)
+
+    volume_list = modal.Volume.objects.list(client=client)
+    assert len(volume_list) == 5
+    assert all(v.name.startswith("test-volume-") for v in volume_list)
+    assert all(v.info().created_by == servicer.default_username for v in volume_list)
+
+    volume_list = modal.Volume.objects.list(max_objects=2, client=client)
+    assert len(volume_list) == 2