+++ CHANGELOG.md
@@ -6,6 +6,17 @@ This changelog documents user-facing updates (features, enhancements, fixes, and
 
 <!-- NEW CONTENT GENERATED BELOW. PLEASE PRESERVE THIS COMMENT. -->
 
+#### 1.1.2.dev16 (2025-08-06)
+
+- Added `.objects.delete()` methods
+- Deprecated existing `.delete()` static methods due to risks of confusion with content deletion operations.
+
+
+#### 1.1.2.dev12 (2025-08-06)
+
+- Added a `.options` namespace to Modal resource object types (`modal.Dict`, `modal.Queue`, `modal.Volume`, and `modal.Secret`).
+
+
 #### 1.1.2.dev3 (2025-08-04)
 
 - Fixed a bug that would cause e.g. some image builds to fail with `'FilePatternMatcher' object has no attribute 'patterns'` when using a `FilePatternMatcher.from_file` ignore pattern.